single ImageView implementation

+++ crates/chess-corners/src/image.rs
@@ -1,8 +1,8 @@
 //! Optional `image::GrayImage` helpers for the unified corner detector.
 
 use crate::multiscale::{find_chess_corners, find_chess_corners_with_refiner};
-use crate::pyramid::ImageView;
 use crate::{ChessConfig, CornerDescriptor, RefinerKind};
+use chess_corners_core::ImageView;
 use image::GrayImage;
 
 /// Detect chessboard corners from a `GrayImage`.
@@ -13,8 +13,8 @@ use image::GrayImage;
 /// [`CornerDescriptor`] values in full-resolution pixel coordinates.
 #[must_use]
 pub fn find_chess_corners_image(img: &GrayImage, cfg: &ChessConfig) -> Vec<CornerDescriptor> {
-    let view =
-        ImageView::from_u8_slice(img.width(), img.height(), img.as_raw()).expect("valid view");
+    let view = ImageView::from_u8_slice(img.width() as usize, img.height() as usize, img.as_raw())
+        .expect("valid view");
     find_chess_corners(view, cfg)
 }
 
@@ -25,7 +25,7 @@ pub fn find_chess_corners_image_with_refiner(
     cfg: &ChessConfig,
     refiner: &RefinerKind,
 ) -> Vec<CornerDescriptor> {
-    let view =
-        ImageView::from_u8_slice(img.width(), img.height(), img.as_raw()).expect("valid view");
+    let view = ImageView::from_u8_slice(img.width() as usize, img.height() as usize, img.as_raw())
+        .expect("valid view");
     find_chess_corners_with_refiner(view, cfg, refiner)
 }
